Clear buffer after dumping the line
Our previous RFID readers prefixed their input with "LEFT SHIFT", which
was the sign to the script to clear the input. We have new now, and they
don't prefix their input. Thus, we will clear the buffer on "ENTER",
right after dumping the formatted line in the CSV file.

While here, do a bit of tidy-up:
- group constants together
- don't include Revdev in the global namespace