Add config for networks
It sets the interface to use DHCP without "auto", so that boot time is
not affected.

The main instance may still need the static config though, due to the
fact it has to be mounted on the Excel host computer, and it is more
convenient for it to expect a fixed IP address.